style: correct type hints about list, rename myClasses to types
