+++ babkin_v/task03/src/modification.py
@@ -1,4 +1,7 @@
-def modificateFile(elems: list):
+from types import Node, Frame
+
+
+def modificateFile(elems: list[list[Node], list[Frame]]):
     frames = elems[1]
     for frame in frames:
         frame.bones[0].position.to_zero()
